--- a/src/server/cluster/cluster_family.cc
+++ b/src/server/cluster/cluster_family.cc
@@ -977,6 +977,15 @@ void ClusterFamily::DflyMigrateFlow(CmdArgList args, SinkReplyBuilder* builder,
 
   builder->SendOk();
 
+  // Try migrating the connection if we have the same shard configuration
+  if (migration->ShardNum() == shard_set->size()) {
+    DCHECK_LT(shard_id, shard_set->size());
+    if (bool success = cntx->conn()->Migrate(shard_set->pool()->at(shard_id)); !success) {
+      builder->SendError("invalid state");
+      return;
+    }
+  }
+
   migration->StartFlow(shard_id, cntx->conn()->socket());
 }