Enable migrations for shard flows on destination instance.

Benchmarking: 1M small keys 
||before|after|
|-|---|---|
|time|10s|8s|
|cpu (dest)| 100/400% | 60%/400%|

Note that increasing destination throughput increases source node cpu usage